Separate DDL scripts for PG < 10 and > 10
For PG >= 10.0, use HASH index, which is now
WAL safe.
